+++ src/eventhandlers/uinputeventhandler.cpp
@@ -231,7 +231,7 @@ void UInputEventHandler::sendMouseButtonEvent(JoyButtonSlot *slot, bool pressed)
                 }
             }
 
-            write_uinput_event(mouseFileHandler, EV_KEY, static_cast<int>(tempcode), pressed ? 1 : 0);
+            write_uinput_event(mouseFileHandler, EV_KEY, tempcode, pressed ? 1 : 0);
         }
         else if (code == 4)
         {
@@ -298,8 +298,8 @@ void UInputEventHandler::sendMouseSpringEvent(int xDis, int yDis,
         double midwidth = width / 2.0;
         double midheight = height / 2.0;
 
-        int fx = static_cast<int>(ceil(32767 * ((xDis - midwidth) / midwidth)));
-        int fy = static_cast<int>(ceil(32767 * ((yDis - midheight) / midheight)));
+        int fx = ceil(32767 * ((xDis - midwidth) / midwidth));
+        int fy = ceil(32767 * ((yDis - midheight) / midheight));
         sendMouseAbsEvent(fx, fy, -1);
     }
 }
@@ -310,8 +310,8 @@ void UInputEventHandler::sendMouseSpringEvent(int xDis, int yDis)
     if ((xDis >= -1.0) && (xDis <= 1.0) &&
         (yDis >= -1.0) && (yDis <= 1.0))
     {
-        int fx = static_cast<int>(ceil(32767 * xDis));
-        int fy = static_cast<int>(ceil(32767 * yDis));
+        int fx = ceil(32767 * xDis);
+        int fy = ceil(32767 * yDis);
         sendMouseAbsEvent(fx, fy, -1);
     }
 }
@@ -499,8 +499,8 @@ void UInputEventHandler::write_uinput_event(int filehandle, int type,
 
     memset(&ev, 0, sizeof(struct input_event));
     gettimeofday(&ev.time, nullptr);
-    ev.type = static_cast<unsigned short>(type);
-    ev.code = static_cast<unsigned short>(code);
+    ev.type = type;
+    ev.code = code;
     ev.value = value;
 
     write(filehandle, &ev, sizeof(struct input_event));
@@ -574,11 +574,11 @@ void UInputEventHandler::sendTextEntryEvent(QString maintext)
             {
                 QtX11KeyMapper::charKeyInformation tempX11 = nativeWinKeyMapper->getCharKeyInformation(maintext.at(i));
                 tempX11.virtualkey = X11Extras::getInstance()->getGroup1KeySym(tempX11.virtualkey);
-                unsigned int tempQtKey = static_cast<unsigned int>(nativeWinKeyMapper->returnQtKey(tempX11.virtualkey));
+                unsigned int tempQtKey = nativeWinKeyMapper->returnQtKey(tempX11.virtualkey);
 
                 if (tempQtKey > 0)
                 {
-                    temp.virtualkey = keymapper->returnVirtualKey(static_cast<int>(tempQtKey));
+                    temp.virtualkey = keymapper->returnVirtualKey(tempQtKey);
                     temp.modifiers = tempX11.modifiers;
                 }
                 else
@@ -601,7 +601,7 @@ void UInputEventHandler::sendTextEntryEvent(QString maintext)
                     testAndAppend(temp.modifiers.testFlag(Qt::MetaModifier), tempList, KEY_LEFTMETA);
                 }
 
-                tempList.append(static_cast<unsigned int>(temp.virtualkey));
+                tempList.append(temp.virtualkey);
                 write_uinput_event(keyboardFileHandler, EV_KEY, temp.virtualkey, 1, true);
             }
 
@@ -614,7 +614,7 @@ void UInputEventHandler::sendTextEntryEvent(QString maintext)
                 {
                     unsigned int currentcode = tempiter.previous();
                     bool sync = !tempiter.hasPrevious() ? true : false;
-                    write_uinput_event(keyboardFileHandler, EV_KEY, static_cast<int>(currentcode), 0, sync);
+                    write_uinput_event(keyboardFileHandler, EV_KEY, currentcode, 0, sync);
                 }
             }
         }
@@ -627,7 +627,7 @@ void UInputEventHandler::testAndAppend(bool tested, QList<unsigned int>& tempLis
     if (tested)
     {
         tempList.append(key);
-        write_uinput_event(keyboardFileHandler, EV_KEY, static_cast<int>(key), 1, false);
+        write_uinput_event(keyboardFileHandler, EV_KEY, key, 1, false);
     }
 }
